refactor: split large files — dispatch, executor, inbox monitor
- autonomous_dispatch.py (998→68 LOC): Split AutonomousCliApprovalGate
  into approval_gate.py and AutonomousDispatchRouter into
  dispatch_router.py. Data classes and re-exports stay in original
  module for backward compatibility.

- executor/engine.py (886→644 LOC): Extract StepDispatcher class into
  step_dispatcher.py. Isolates step-level dispatch logic (CC invocation,
  API/CLI routing, workaround recovery) from task-level state machine.

- inbox/monitor.py: Decompose 875-line _check_once_inner() into 60-line
  orchestrator + 4 phase methods (_phase_resume, _phase_detect_changes,
  _phase_create_records, _phase_dispatch_batches). Same file, no new
  classes — method extraction only.

All existing tests pass (5573 passed). Zero regressions — 19 pre-existing
failures in hooks/outreach confirmed on base branch.
